convert周りのリファクタリング
