+++ application/api/util/converter/converter.go
@@ -101,189 +101,6 @@ func RequestAlarmToDatabaseAlarm(alarm request.Alarm, target string) database.Al
 	return databaseAlarm
 }
 
-func DatabaseCharaListToResponseCharaList(charaList []database.Chara, baseURL string) []response2.Chara {
-	responseCharaList := make([]response2.Chara, 0)
-	for i := 0; i < len(charaList); i++ {
-		responseChara := DatabaseCharaToResponseChara(charaList[i], baseURL)
-		responseCharaList = append(responseCharaList, responseChara)
-	}
-	return responseCharaList
-}
-
-func DatabaseCharaToResponseChara(databaseChara database.Chara, baseURL string) response2.Chara {
-	return response2.Chara{
-		CharaID:     databaseChara.CharaID,
-		Enable:      databaseChara.Enable,
-		Name:        databaseChara.Name,
-		CreatedAt:   databaseChara.CreatedAt,
-		UpdatedAt:   databaseChara.UpdatedAd,
-		Description: databaseChara.Description,
-		Profiles:    databaseCharaProfileListToResponseCharaProfileList(databaseChara.Profiles),
-		Resources:   databaseCharaToResponseCharaResourceList(databaseChara, baseURL),
-		Expression:  databaseCharaExpressionMapToResponseCharaExpressionMap(databaseChara.Expressions, baseURL, databaseChara.CharaID),
-		Calls:       databaseCharaCallListToResponseCharaCallList(databaseChara.Calls, baseURL, databaseChara.CharaID),
-	}
-}
-
-func databaseCharaProfileListToResponseCharaProfileList(databaseCharaProfileList []database.CharaProfile) []response2.CharaProfile {
-	responseCharaProfileList := make([]response2.CharaProfile, 0)
-	for i := 0; i < len(databaseCharaProfileList); i++ {
-		responseCharaProfile := databaseCharaProfileToResponseCharaProfile(databaseCharaProfileList[i])
-		responseCharaProfileList = append(responseCharaProfileList, responseCharaProfile)
-	}
-	return responseCharaProfileList
-}
-
-func databaseCharaProfileToResponseCharaProfile(databaseCharaProfile database.CharaProfile) response2.CharaProfile {
-	return response2.CharaProfile{
-		Title: databaseCharaProfile.Title,
-		Name:  databaseCharaProfile.Name,
-		URL:   databaseCharaProfile.URL,
-	}
-}
-
-func databaseCharaToResponseCharaResourceList(databaseChara database.Chara, resourceBaseURL string) []response2.CharaResource {
-	// response2.CharaResourcesを作成
-	responseCharaResources := make([]response2.CharaResource, 0)
-
-	// expressionsのリソースを生成
-	for _, databaseCharaExpression := range databaseChara.Expressions {
-		for _, imageFileName := range databaseCharaExpression.ImageFileNames {
-			responseCharaResources = append(responseCharaResources, response2.CharaResource{
-				FileURL: createFileURL(resourceBaseURL, databaseChara.CharaID, imageFileName),
-			})
-		}
-
-		for _, voiceFileName := range databaseCharaExpression.VoiceFileNames {
-			responseCharaResources = append(responseCharaResources, response2.CharaResource{
-				FileURL: createFileURL(resourceBaseURL, databaseChara.CharaID, voiceFileName),
-			})
-		}
-	}
-
-	// callsのリソースを生成
-	for _, databaseCharaCall := range databaseChara.Calls {
-		responseCharaResources = append(responseCharaResources, response2.CharaResource{
-			FileURL: createFileURL(resourceBaseURL, databaseChara.CharaID, databaseCharaCall.VoiceFileName),
-		})
-	}
-
-	// TODO: responseCharaResources の中から重複要素を削除
-	return responseCharaResources
-}
-
-func databaseCharaExpressionMapToResponseCharaExpressionMap(databaseCharaExpressionMap map[string]database.CharaExpression, baseURL string, charaID string) map[string]response2.CharaExpression {
-	responseCharaExpressionMap := map[string]response2.CharaExpression{}
-	for key, databaseCharaExpression := range databaseCharaExpressionMap {
-		// 画像とボイスにBase URLを追加する
-		responseImages := make([]string, 0)
-		for _, imageFileName := range databaseCharaExpression.ImageFileNames {
-			responseImages = append(responseImages, createFileURL(baseURL, charaID, imageFileName))
-		}
-		responseVoices := make([]string, 0)
-		for _, voiceFileName := range databaseCharaExpression.VoiceFileNames {
-			responseVoices = append(responseVoices, createFileURL(baseURL, charaID, voiceFileName))
-		}
-
-		responseCharaExpression := response2.CharaExpression{
-			ImageFileURLs: responseImages,
-			VoiceFileURLs: responseVoices,
-		}
-		responseCharaExpressionMap[key] = responseCharaExpression
-	}
-	return responseCharaExpressionMap
-}
-
-func databaseCharaCallListToResponseCharaCallList(databaseCharaCallList []database.CharaCall, baseURL string, charaID string) []response2.CharaCall {
-	responseCharaCallList := make([]response2.CharaCall, 0)
-	for i := 0; i < len(databaseCharaCallList); i++ {
-		responseCharaCall := databaseCharaCallToResponseCharaCall(databaseCharaCallList[i], baseURL, charaID)
-		responseCharaCallList = append(responseCharaCallList, responseCharaCall)
-	}
-	return responseCharaCallList
-}
-
-func databaseCharaCallToResponseCharaCall(databaseCharaCall database.CharaCall, baseURL string, charaID string) response2.CharaCall {
-	return response2.CharaCall{
-		Message:       databaseCharaCall.Message,
-		VoiceFileName: databaseCharaCall.VoiceFileName,
-		VoiceFileURL:  createFileURL(baseURL, charaID, databaseCharaCall.VoiceFileName),
-	}
-}
-
-func DatabaseAlarmToResponseAlarm(alarm database.Alarm) response2.Alarm {
-	// UTCのdatabase.Alarmを時差のあるresponse.Alarmに変換する
-	var alarmHour int
-	var alarmMinute int
-	var alarmSunday bool
-	var alarmMonday bool
-	var alarmTuesday bool
-	var alarmWednesday bool
-	var alarmThursday bool
-	var alarmFriday bool
-	var alarmSaturday bool
-
-	// 時差を計算
-	diff := (float32(alarm.Hour) + float32(alarm.Minute)/60.0) + alarm.TimeDifference
-	if diff > 24 {
-		// tomorrow
-		diff -= 24.0
-		alarmHour = int(diff)
-		alarmMinute = int((diff-float32(alarmHour))*60 + 0.5)
-		alarmSunday = alarm.Monday
-		alarmMonday = alarm.Tuesday
-		alarmTuesday = alarm.Wednesday
-		alarmWednesday = alarm.Thursday
-		alarmThursday = alarm.Friday
-		alarmFriday = alarm.Saturday
-		alarmSaturday = alarm.Sunday
-	} else if diff >= 0 {
-		// today
-		alarmHour = int(diff)
-		alarmMinute = int((diff-float32(alarmHour))*60 + 0.5)
-		alarmSunday = alarm.Sunday
-		alarmMonday = alarm.Monday
-		alarmTuesday = alarm.Tuesday
-		alarmWednesday = alarm.Wednesday
-		alarmThursday = alarm.Thursday
-		alarmFriday = alarm.Friday
-		alarmSaturday = alarm.Saturday
-	} else {
-		// yesterday
-		diff += 24.0
-		alarmHour = int(diff)
-		alarmMinute = int((diff-float32(alarmHour))*60 + 0.5)
-		alarmSaturday = alarm.Friday
-		alarmFriday = alarm.Thursday
-		alarmThursday = alarm.Wednesday
-		alarmWednesday = alarm.Tuesday
-		alarmTuesday = alarm.Monday
-		alarmMonday = alarm.Sunday
-		alarmSunday = alarm.Saturday
-	}
-
-	return response2.Alarm{
-		AlarmID:        alarm.AlarmID,
-		UserID:         alarm.UserID,
-		Type:           alarm.Type,
-		Enable:         alarm.Enable,
-		Name:           alarm.Name,
-		Hour:           alarmHour,
-		Minute:         alarmMinute,
-		TimeDifference: alarm.TimeDifference,
-		CharaID:        alarm.CharaID,
-		CharaName:      alarm.CharaName,
-		VoiceFileName:  alarm.VoiceFileName,
-		Sunday:         alarmSunday,
-		Monday:         alarmMonday,
-		Tuesday:        alarmTuesday,
-		Wednesday:      alarmWednesday,
-		Thursday:       alarmThursday,
-		Friday:         alarmFriday,
-		Saturday:       alarmSaturday,
-	}
-}
-
 // 文字を*に変換
 func maskAuthToken(authToken string) string {
 	length := len(authToken)
@@ -298,8 +115,4 @@ func maskAuthToken(authToken string) string {
 		}
 	}
 	return r
-}
-
-func createFileURL(resourceBaseURL string, charaID string, fileName string) string {
-	return resourceBaseURL + "/" + charaID + "/" + fileName
 }
